+++ firebase-config/src/test/java/com/google/firebase/remoteconfig/RemoteConfigTests.kt
@@ -29,8 +29,8 @@ import com.google.firebase.platforminfo.UserAgentPublisher
 import com.google.firebase.remoteconfig.internal.ConfigCacheClient
 import com.google.firebase.remoteconfig.internal.ConfigFetchHandler
 import com.google.firebase.remoteconfig.internal.ConfigGetParameterHandler
-import com.google.firebase.remoteconfig.internal.ConfigMetadataClient
 import com.google.firebase.remoteconfig.internal.ConfigRealtimeHandler
+import com.google.firebase.remoteconfig.internal.ConfigSharedPrefsClient
 import com.google.firebase.remoteconfig.internal.rollouts.RolloutsStateSubscriptionsHandler
 import org.junit.After
 import org.junit.Before
@@ -139,13 +139,26 @@ class ConfigTests : BaseTestCase() {
         defaultConfigsCache = mock(ConfigCacheClient::class.java),
         fetchHandler = mock(ConfigFetchHandler::class.java),
         getHandler = mockGetHandler,
-        frcMetadata = mock(ConfigMetadataClient::class.java),
+        frcSharedPrefs = mock(ConfigSharedPrefsClient::class.java),
         realtimeHandler = mock(ConfigRealtimeHandler::class.java),
         rolloutsStateSubscriptionsHandler = mock(RolloutsStateSubscriptionsHandler::class.java)
       )
 
     `when`(mockGetHandler.getValue("KEY")).thenReturn(StringRemoteConfigValue("non default value"))
     assertThat(remoteConfig["KEY"].asString()).isEqualTo("non default value")
+  }
+
+  @Test
+  fun `Custom Signals builder support multiple types`() {
+    val customSignals = customSignals {
+      put("key1", "value1")
+      put("key2", 123L)
+      put("key3", 45.67)
+      put("key4", null)
+    }
+    val expectedSignals =
+      mapOf("key1" to "value1", "key2" to "123", "key3" to "45.67", "key4" to null)
+    assertThat(customSignals.customSignals).isEqualTo(expectedSignals)
   }
 }