+++ firebase-vertexai/src/main/kotlin/com/google/firebase/vertexai/type/Exceptions.kt
@@ -18,7 +18,6 @@ package com.google.firebase.vertexai.type
 
 import com.google.firebase.vertexai.FirebaseVertexAI
 import com.google.firebase.vertexai.common.FirebaseCommonAIException
-import com.google.firebase.vertexai.internal.util.toPublic
 import kotlinx.coroutines.TimeoutCancellationException
 
 /** Parent class for any errors that occur from the [FirebaseVertexAI] SDK. */
@@ -45,7 +44,7 @@ internal constructor(message: String, cause: Throwable? = null) : RuntimeExcepti
             is com.google.firebase.vertexai.common.InvalidAPIKeyException ->
               InvalidAPIKeyException(cause.message ?: "")
             is com.google.firebase.vertexai.common.PromptBlockedException ->
-              PromptBlockedException(cause.response.toPublic(), cause.cause)
+              PromptBlockedException(cause.response?.toPublic(), cause.cause)
             is com.google.firebase.vertexai.common.UnsupportedUserLocationException ->
               UnsupportedUserLocationException(cause.cause)
             is com.google.firebase.vertexai.common.InvalidStateException ->
@@ -58,6 +57,8 @@ internal constructor(message: String, cause: Throwable? = null) : RuntimeExcepti
               ServiceDisabledException(cause.message ?: "", cause.cause)
             is com.google.firebase.vertexai.common.UnknownException ->
               UnknownException(cause.message ?: "", cause.cause)
+            is com.google.firebase.vertexai.common.ContentBlockedException ->
+              ContentBlockedException(cause.message ?: "", cause.cause)
             else -> UnknownException(cause.message ?: "", cause)
           }
         is TimeoutCancellationException ->
@@ -88,13 +89,22 @@ internal constructor(message: String, cause: Throwable? = null) :
  *
  * @property response The full server response.
  */
-// TODO(rlazo): Add secondary constructor to pass through the message?
 public class PromptBlockedException
-internal constructor(public val response: GenerateContentResponse, cause: Throwable? = null) :
+internal constructor(
+  public val response: GenerateContentResponse?,
+  cause: Throwable? = null,
+  message: String? = null,
+) :
   FirebaseVertexAIException(
-    "Prompt was blocked: ${response.promptFeedback?.blockReason?.name}",
+    "Prompt was blocked: ${response?.promptFeedback?.blockReason?.name?: message}",
     cause,
-  )
+  ) {
+  internal constructor(message: String, cause: Throwable? = null) : this(null, cause, message)
+}
+
+public class ContentBlockedException
+internal constructor(message: String, cause: Throwable? = null) :
+  FirebaseVertexAIException(message, cause)
 
 /**
  * The user's location (region) is not supported by the API.